Add enum/static const and other testing
This verifies that #1282 is fixed by specifying a dialog id and then using it as
a static const int id.

I also added a french title to verify UTF8 conversion, changed the static text
variable to public: access and added a comment, and added a custom
member with a comment.