Start timing clock sooner
Previously the internal timer only dealt with form code generation. This commit
moves it near the top. Now the the time to create any .cmake file is tracked.
It also will include the time in the new Project::FindWxueFunctions() used
to figure out which form needs to write the single copy of one or more of the
wxueFunctions.