Special-case no_closing_brace
This changes how a C++ header file is generated when the no_closing_brace
property is set to true. When a new file is generated, the // clang format on
line will be removed from within the comment block, and placed after the
comment block instead. It will also generate the initial closing brace so that
the source file can be immediately compiled before the user modifies the
header file.