Update build notes
The original plan was to maintain a wxUiEditor_1_2 branch that would contain
only changes important to a 1.2.1 release. However, there are so many changes
that should be in the next release, that it became fairly difficult to figure out
which PRs/commits should be cherry-picked into that branch. This is being
replaced with a wxUiEditor_1_2_1 branch that is rebased from main.