Change algorithm converting filename to varname
A completely UTF8 filename could triple it's size since very UTF8 character was
preceeded by an '_'. In addition, a completely UTF8 filename would not convert
anything.

The new algorithm uses snprintf to convert to a hex digit, and only adds a
preceeding '_' character if the preceeding filename character was *not* a
UTF8 character.