Modify AddEmbeddedBundleImage, add get_bundle
With this commit, all the images that are part of a bundle are added to the
embed->imgs vector, and embed->get_bundle() will load the images and
create the bundle. This makes it possible to get wxBitmapBundles as needed
without needing to store anything more than we normally store -- i.e., the
compressed image data (png, svg+zlib, etc).